Fix build in C++0x mode
This fixes compiler errors (gcc 4.6 -std=gnu++0x on x86_64 Linux):

embedded/qwslock.cpp: In function `bool forceLock(int, int, int)':
embedded/qwslock.cpp:121:39: error: narrowing conversion of `semNum'
        from `int' to `short unsigned int' inside { } [-fpermissive]

(and equivalent errors in other lines/files)

See
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2011/n3242.pdf
Section 8.5.4/6

Merge-Request: 1240
Reviewed-on: http://codereview.qt.nokia.com/275

+++ src/gui/embedded/qwslock.cpp
@@ -114,7 +114,7 @@ QWSLock::~QWSLock()
     QWSSignalHandler::instance()->removeSemaphore(semId);
 }
 
-static bool forceLock(int semId, int semNum, int)
+static bool forceLock(int semId, unsigned short semNum, int)
 {
     int ret;
     do {
@@ -135,7 +135,7 @@ static bool forceLock(int semId, int semNum, int)
     return (ret != -1);
 }
 
-static bool up(int semId, int semNum)
+static bool up(int semId, unsigned short semNum)
 {
     int ret;
     do {
@@ -148,7 +148,7 @@ static bool up(int semId, int semNum)
     return (ret != -1);
 }
 
-static bool down(int semId, int semNum)
+static bool down(int semId, unsigned short semNum)
 {
     int ret;
     do {
@@ -161,7 +161,7 @@ static bool down(int semId, int semNum)
     return (ret != -1);
 }
 
-static int getValue(int semId, int semNum)
+static int getValue(int semId, unsigned short semNum)
 {
     int ret;
     do {
@@ -210,7 +210,7 @@ void QWSLock::unlock(LockType type)
             return;
     }
 
-    const int semNum = type;
+    const unsigned short semNum = type;
     int ret;
     do {
         sembuf sops = {semNum, 1, 0};
